// server/src/ts/interfaces/customer.interfaces.ts
export interface registerData {
    firstname: string;
    lastname: string;
    email: string;
    password: string;
    birthdate: string;
    img: string;
    gender: string;
}

export interface userData {
    firstname: string;
    lastname: string;
    email: string;
    birthdate: string;
    firstconnect: boolean;
    firstcircle: boolean;
    img: string;
    created_at: string;
    updated_at: string;
}